+++ frontend/src/widgets/FrequencyGraph.css
@@ -1,0 +1,52 @@
+.ttahub--frequency-graph {
+  width: 100%;
+}
+
+@media(min-width: 1090px){
+  .ttahub--show-accessible-data-button {
+    flex: 1;
+    text-align: right;
+  }
+}
+
+.ttahub--frequency-graph-tooltip {
+  background: black;
+  border-radius: 5px;
+  color: white;
+  min-width: 2rem;
+  padding: .5em;
+  position: fixed;
+  text-align: center;
+  z-index: 2;
+}
+
+.ttahub--frequency-graph-tooltip::after {
+  border-top: 5px solid black;
+  border-bottom: 5px solid transparent;
+  border-left: 5px solid transparent;
+  border-right: 5px solid transparent;
+  height: 0;
+  width: 0;
+  content: '';
+  position: absolute;
+  top: 100%;
+  left: 33%;
+}
+
+
+.tta-dashboard--frequency-graph-apply-filters {
+  height: 40px;
+  min-width: 120px;
+}
+
+@media(max-width: 768px) {
+  .ttahub--frequency-graph-control-row {
+      display: block;
+      width: 100%;
+  }
+
+  .ttahub--frequency-graph-control-row .usa-select,
+  .tta-dashboard--frequency-graph-apply-filters {
+      width: 180px;
+  }
+}
